Move XCTestCase extension to a separate file.
